+++ JWTweak.py
@@ -1,4 +1,5 @@
 #JWTweak-v1.1.1
+
 import os
 import jwt
 import re